+++ listui/src/widgets/loading.rs
@@ -1,7 +1,4 @@
-use std::io::Stdout;
-
 use ratatui::Frame;
-use ratatui::backend::CrosstermBackend;
 use ratatui::layout::{Rect, Layout, Constraint, Alignment};
 use ratatui::style::Style;
 use ratatui::widgets::Paragraph;
@@ -22,8 +19,6 @@ r"
 ;;;;;;            ;;;;;;      
 `;;;;'            `;;;;'      ";
 
-static SPINNER_FRAMES: [char; 8] = ['|', '/', '―', '\\', '|', '/', '―', '\\'];
-
 pub struct LoadingWidget {
     label: String,
     frame: u16,
@@ -34,12 +29,16 @@ impl LoadingWidget {
     pub fn new(label: &str) -> Self {
 
         Self {
-            label: format!("{} {}", label, SPINNER_FRAMES[0]),
+            label: label.to_string(),
             frame: 0
         }
     }
 
-    pub fn draw(&mut self, frame: &mut Frame<CrosstermBackend<Stdout>>, area: Rect) {
+    pub fn change_label(&mut self, label: String) {
+        self.label = label;
+    }
+
+    pub fn draw(&mut self, frame: &mut Frame, area: Rect) {
         
         if area.height < 20 {
             frame.render_widget(Paragraph::new(self.label.as_str()).style(Style::default().fg(super::ACC_COLOR)).alignment(Alignment::Center), area);  
@@ -52,8 +51,6 @@ impl LoadingWidget {
                 .margin(1)
                 .split(area);
             
-            self.label.pop();
-            self.label.push(SPINNER_FRAMES[self.frame as usize]);
             self.frame = (self.frame + 1) % 8;
             frame.render_widget(Paragraph::new(self.label.as_str()).style(Style::default().fg(super::ACC_COLOR)).alignment(Alignment::Center), chunks[0]);
             frame.render_widget(Paragraph::new(FIGURE).style(Style::default().fg(super::ACC_COLOR)).alignment(Alignment::Center), chunks[1]);